# Improve performance while caching materializers