--- a/lib/mini_sql/postgres/deserializer_cache.rb
+++ b/lib/mini_sql/postgres/deserializer_cache.rb
@@ -14,13 +14,12 @@ module MiniSql
       def materializer(result)
         key = result.fields
 
-        materializer = @cache.delete(key)
-        if materializer
-          @cache[key] = materializer
-        else
-          materializer = @cache[key] = new_row_matrializer(result)
-          @cache.shift if @cache.length > @max_size
-        end
+        materializer = @cache[key] ||
+          begin
+            _materializer = @cache[key] = new_row_matrializer(result)
+            @cache.shift if @cache.length > @max_size
+            _materializer
+          end
 
         materializer
       end
@@ -31,13 +30,12 @@ module MiniSql
         key = result.fields
 
         # trivial fast LRU implementation
-        materializer = @cache.delete(key)
-        if materializer
-          @cache[key] = materializer
-        else
-          materializer = @cache[key] = new_row_matrializer(result)
-          @cache.shift if @cache.length > @max_size
-        end
+        materializer = @cache[key] ||
+          begin
+            _materializer = @cache[key] = new_row_matrializer(result)
+            @cache.shift if @cache.length > @max_size
+            _materializer
+          end
 
         if decorator_module
           materializer = materializer.decorated(decorator_module)