With this small change we gain a bit performance. See the results of a sandbox (for a happy path only) profiling:

```ruby
require 'benchmark/ips'

@cache = {
  x: 1,
  y: 2,
  z: 3
}

GC.disable

def old_call
  key = :x
  m = @cache.delete(key)
  @cache[key] = m
  m
end

def new_call
  key = :x
  m = @cache[key] || 0
  m
end

Benchmark.ips do |x|
  x.report(:old) do
    old_call
  end
  x.report(:new) do
    new_call
  end
  x.compare!
end
```

```
Warming up --------------------------------------
                 old   685.254k i/100ms
                 new     1.070M i/100ms
Calculating -------------------------------------
                 old      6.721M (± 2.9%) i/s -     33.577M in   5.000392s
                 new     10.486M (± 1.1%) i/s -     52.426M in   5.000098s

Comparison:
                 new: 10486303.5 i/s
                 old:  6721441.8 i/s - 1.56x  (± 0.00) slower
```