@xronos-i-am 2021-02-19
You loose LRU implementation. The most used materializer may be added to cache first. New materializers added to the end of cache. At the some moment cache will exceed `max_size` and first materializer will shifted from cache.